+++ index.html
@@ -29,7 +29,7 @@
     </script>
   </head>
   <body>
-    <script src="main.js"></script>
+    <script src="src/index.js" type="module"></script>
     <script>
       (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
       (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
