+++ src/gamedata.js
@@ -1,0 +1,8 @@
+var gameData = {
+  level: -1,
+  gameOver: false,
+  healthPoints: 100,
+  gemPoints: 0
+};
+
+export default { gameData };
